Adding errors Request/Response reads
Currently in order to read from a request or response we needed to read
the entire body into a buffer and then read from that buffer. This
approach does not return any errors that were encountered while reading
from the network. Instead errors are written into the Request/Response
body and can be read from there.

This makes it very hard to reliably determine if an error has occurred.
I have added two very basic functions which do return an explicit error.

One implements current behaviour, read body into buffer and return that,
but returns an explicit error. The second exposes the bodyStream field
as an io.Reader allowing client code to stream data with minimal memory
consumption, as well as exposing errors.