Optimize deletion of arg from []argsKV
The previous behavior would perform a deletion by making a temporary
copy of the arg to be deleted, reindexing the remainder of the array to
shift it back to the position of the removed item, copying the temporary
copy to the current last index, and then shrinking the length by one.

This change simplifies the operation by simply copying the last arg to
the position of the removal, and then shrinking the length.

Note that this does cause there to be a duplicate arg in the array,
which is beyond the length, but within the capacity. This should only
matter if the array has its length increased without updating the newly
exposed arg. I don't believe this happens.

If the duplicate item could be an issue, the same approach could be
taken but with a 2 way swap.

```
args[i], args[n] = args[n], args[i]
```