Fix client.go.
Fix the bug that not setting read deadline in a new connection caused by old lastReadDeadlineTime.
You have checked the cc.lastReadDeadlineTime in doNonNilReqResp(), it's good to improve performance. But, you should reset all the fields of clientConn when you release it. Or the new connection may use the lastReadDeadlineTime of the old connection to decide if it should set the read deadline.