Stop random TestServerErrSmallBuffer failures
TestServerErrSmallBuffer had a small race condition where the test would
checkout the output of the logger after the client connection had been
served but before the worker had written the result to the logger.

This new code is also much faster as it doesn't actually use TCP but
just some in memory buffers.