Add ResponseErrorHandler for errors while writing the response
The handler is called if there was an error writing to the TCP connection.

Business-critical operations require to store the final result sent to the
client. For cloud providers's load balancer the problem is more subtle,
the final result sent to the customer may be different from the application
so it's important to log that the response was properly read (or discarded)
by the LB.