Add ServerTrace
This adds a mechanism to trace events that occur within the Server. I
added hooks for some common use cases that I could think of but it can
easily be expanded to other things as people see fit. The pattern is
heavily inspired by
[`net/http/httptrace`](https://golang.org/pkg/net/http/httptrace/#ClientTrace).

The most common use case for this will be logging and metrics: it allows
centrally logging the status code that was sent to clients, bytes
transferred in the process, errors that might have occured during the
transfer, etc.

fasthttp already has some limited tracing capability for connection
state which I now marked as deprecated since the new tracing mechanism
encompasses all the previous hooks.

A side note regarding backward compatibility: I ended up adding an
additional int64 return value for many of the existing Write methods.
This was necessary in order to provide bytes written to the respective
trace hook - which I think is a very valuable data point for logging and
stats that might get generated. This however breaks backward
compatibility but it is trivial to update as the additional return
argument can simply be ignored.