uri.go fast scheme comparison
The methods isHttps() and isHttp() checks that all bytes are equals bytes.Equal().
If the URI is reused for many requests then the bytes.Equal() call is performed on each request.
Another downside is that on SetScheme() memory is once allocated and copied.
FastHttp supports only http and https so instead we may store a strHTTP(S) constant and then just compare by reference.
But in Go byte slices doesn't have doesn't have an identity so we must use a pointer to a byte slice instead.

Also for a custom schemes (ftp, dav) the memory will be allocated instead of pointing to a constant. If really needed this may be changed in the future.