uri.go store host with port
for a faster lookup in the client the HostWithPort method returns just a host field.
The host field is always contains both hostname and port even for standard ports 80 and 443.
For a backward compatibility the Host() method returns host without a port if the port is standard.

Additionally a Port() method added which return only the port part of url.
Internally it returns a host substring starting from portPos.
To distinguish from custom or standard (added) port the portPos is negative for standard ports.
The portPos uses a signed byte so max host length is 127.